flex 03-flex-header-2 solved
This one was more difficult so I'll give a brief explanation of what I did. I added two more divs, left and right. Then in the css I added display:flex and justify content: space-between to the .header. then in the ul I added the display: flex and put a gap of 20px and removed the margin and padding to remove the excess space that came out in the result. then I created a .left and .right in the css and added display:flex, align-items: center and a 20px gap to both